Separate new helpers test file
